requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
pyyaml>=6.0
python-dotenv>=1.0.0
mediapipe>=0.10.0  # Para análisis postural
psutil>=5.9.0
gputil>=1.4.0
tensorflow>=2.12.0
tensorflow-hub>=0.14.0 